added: BigInt_b.c for testing 128 bit divide_by_n() methods

+++ BigInt.h
@@ -61,7 +61,7 @@ typedef struct BigInt {
 
     uint32_t (*add)(void* a, void* b, void* c, BigIntType type);
     int (*compare)(void* a, void* b, BigIntType type);
-    uint32_t (*multiply_by_n)(void* a, uint32_t n, void* b, BigIntType type);
+    uint32_t (*multiply_by_n)(void* a, uint64_t n, void* b, BigIntType type);
     uint32_t (*subtract)(void* a, void* b, void* c, BigIntType type);
     void (*multiply)(void* a, void* b, void* c, BigIntType type);
     uint32_t (*add_by_n)(void* a, uint32_t n, void* b, BigIntType type);
